Add a policy to only allow 4 concurrent build actions
The plugging of github webhook made the unbounded build actions
overrun the openstack quota (for disk volumes total size).

This enforces a limit of 4 concurrent builds which won't be too
heavy on our current openstack infrastructure. This should be a
safe baseline.

The subsequent requested build actions will be queued until one
finishes.